+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 - **InfluxDB:** Added capability to acquire bulk readings in JSON format
 - **MQTT:** Added capability to acquire bulk readings in compact JSON format,
   with timestamps as keys
+- **InfluxDB:** Added decoder for acquiring data in NDJSON format
 
 .. _kotori-0.28.0: